+++ src/__generated__/FairTestsQuery.graphql.ts
@@ -93,6 +93,20 @@ fragment FairHeader_fair on Fair {
     artists
     partners
   }
+  followed_content {
+    artists {
+      name
+      href
+      id
+      _id
+      __id
+    }
+    galleries {
+      _id
+      name
+      __id
+    }
+  }
   partner_names: shows_connection(first: 2) {
     edges {
       node {
@@ -332,6 +346,13 @@ v6 = {
   "storageKey": null
 },
 v7 = [
+  v4,
+  v3,
+  v2,
+  v6,
+  v1
+],
+v8 = [
   {
     "kind": "Literal",
     "name": "first",
@@ -339,45 +360,45 @@ v7 = [
     "type": "Int"
   }
 ],
-v8 = {
+v9 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "__typename",
   "args": null,
   "storageKey": null
 },
-v9 = {
+v10 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "aspect_ratio",
   "args": null,
   "storageKey": null
 },
-v10 = {
+v11 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "url",
   "args": null,
   "storageKey": null
 },
-v11 = {
+v12 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "is_closed",
   "args": null,
   "storageKey": null
 },
-v12 = {
+v13 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "display",
   "args": null,
   "storageKey": null
 },
-v13 = [
-  v12
-],
 v14 = [
+  v13
+],
+v15 = [
   v4,
   v1
 ];
@@ -385,7 +406,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "FairTestsQuery",
-  "id": "f53928f9b1479d08787db8234a15cf10",
+  "id": "34662b47b49122c0e404541e47fcc639",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -514,10 +535,45 @@ return {
           },
           {
             "kind": "LinkedField",
+            "alias": null,
+            "name": "followed_content",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "FollowedContent",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "artists",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "Artist",
+                "plural": true,
+                "selections": v7
+              },
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "galleries",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "Partner",
+                "plural": true,
+                "selections": [
+                  v6,
+                  v4,
+                  v1
+                ]
+              }
+            ]
+          },
+          {
+            "kind": "LinkedField",
             "alias": "partner_names",
             "name": "shows_connection",
             "storageKey": "shows_connection(first:2)",
-            "args": v7,
+            "args": v8,
             "concreteType": "ShowConnection",
             "plural": false,
             "selections": [
@@ -549,7 +605,7 @@ return {
                         "concreteType": null,
                         "plural": false,
                         "selections": [
-                          v8,
+                          v9,
                           v1,
                           {
                             "kind": "InlineFragment",
@@ -586,7 +642,7 @@ return {
             "alias": "artists_names",
             "name": "artists",
             "storageKey": "artists(first:2)",
-            "args": v7,
+            "args": v8,
             "concreteType": "ArtistConnection",
             "plural": false,
             "selections": [
@@ -607,13 +663,7 @@ return {
                     "args": null,
                     "concreteType": "Artist",
                     "plural": false,
-                    "selections": [
-                      v4,
-                      v3,
-                      v2,
-                      v6,
-                      v1
-                    ]
+                    "selections": v7
                   }
                 ]
               }
@@ -635,8 +685,8 @@ return {
                 "args": null,
                 "storageKey": null
               },
-              v9,
-              v10
+              v10,
+              v11
             ]
           },
           v1,
@@ -880,7 +930,7 @@ return {
                         "concreteType": null,
                         "plural": false,
                         "selections": [
-                          v8,
+                          v9,
                           v1,
                           {
                             "kind": "InlineFragment",
@@ -967,7 +1017,7 @@ return {
                                     "concreteType": "Image",
                                     "plural": false,
                                     "selections": [
-                                      v9,
+                                      v10,
                                       {
                                         "kind": "ScalarField",
                                         "alias": null,
@@ -1057,7 +1107,7 @@ return {
                                         "args": null,
                                         "storageKey": null
                                       },
-                                      v11,
+                                      v12,
                                       {
                                         "kind": "ScalarField",
                                         "alias": null,
@@ -1085,7 +1135,7 @@ return {
                                         "args": null,
                                         "concreteType": "SaleArtworkOpeningBid",
                                         "plural": false,
-                                        "selections": v13
+                                        "selections": v14
                                       },
                                       {
                                         "kind": "LinkedField",
@@ -1095,7 +1145,7 @@ return {
                                         "args": null,
                                         "concreteType": "SaleArtworkCurrentBid",
                                         "plural": false,
-                                        "selections": v13
+                                        "selections": v14
                                       },
                                       {
                                         "kind": "ScalarField",
@@ -1113,7 +1163,7 @@ return {
                                         "concreteType": "Sale",
                                         "plural": false,
                                         "selections": [
-                                          v11,
+                                          v12,
                                           v1
                                         ]
                                       },
@@ -1135,7 +1185,7 @@ return {
                                     ],
                                     "concreteType": "Artist",
                                     "plural": true,
-                                    "selections": v14
+                                    "selections": v15
                                   },
                                   {
                                     "kind": "LinkedField",
@@ -1145,7 +1195,7 @@ return {
                                     "args": null,
                                     "concreteType": "Partner",
                                     "plural": false,
-                                    "selections": v14
+                                    "selections": v15
                                   },
                                   v3
                                 ]
@@ -1189,7 +1239,7 @@ return {
                         "args": null,
                         "concreteType": "Fair",
                         "plural": false,
-                        "selections": v14
+                        "selections": v15
                       },
                       {
                         "kind": "LinkedField",
@@ -1200,7 +1250,7 @@ return {
                         "concreteType": "Image",
                         "plural": false,
                         "selections": [
-                          v10
+                          v11
                         ]
                       },
                       {
@@ -1212,12 +1262,12 @@ return {
                         "concreteType": "Location",
                         "plural": false,
                         "selections": [
-                          v12,
+                          v13,
                           v1
                         ]
                       },
                       v1,
-                      v8
+                      v9
                     ]
                   }
                 ]