+++ src/lib/Scenes/City/Components/AllEvents.tsx
@@ -2,11 +2,11 @@ import { Box, Separator, Serif } from "@artsy/palette"
 import { EventSection } from "lib/Scenes/City/Components/EventSection"
 import { BucketKey, BucketResults } from "lib/Scenes/Map/Bucket"
 import React from "react"
-import { FlatList } from "react-native"
+import { FlatList, ViewProperties } from "react-native"
 import { FairEventSection } from "./FairEventSection"
 import { SavedEventSection } from "./SavedEventSection"
 
-interface Props {
+interface Props extends ViewProperties {
   currentBucket: BucketKey
   buckets: BucketResults
   cityName: string